fix: change storybook to typescript
